+++ src/Creator/GitlabPluginProject/CreatorTask.php
@@ -6,7 +6,6 @@ namespace srag\Plugins\SrProjectHelper\Creator\GitlabPluginProject;
 require_once __DIR__ . "/../../../vendor/autoload.php";
 
 use Gitlab\Model\Project;
-use srag\Plugins\SrProjectHelper\Config\Config;
 use srag\Plugins\SrProjectHelper\Creator\Gitlab\AbstractGitlabCreatorTask;
 
 /**
@@ -27,8 +26,8 @@ class CreatorTask extends AbstractGitlabCreatorTask {
 		 */
 		$project = null;
 
-		return $this->getStepsForNewPlugin($data["name"], function (): int {
-			return Config::getField(Config::KEY_GITLAB_PLUGINS_GROUP_ID);
-		}, $data["maintainer_user_id"], $project);
+		return $this->getStepsForNewPlugin($data["name"], function () use (&$data): int {
+			return intval($data["group"]);
+		}, $data["maintainer_user"], $project);
 	}
 }