+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.1.0]
+- Select any group for a plugin project
+- Select maintainer user by name and display an information message, if the user no ILIAS user has created
+
 ## [1.0.1]
 - Protect develop branch